refactor: make imports absolute and separate into components
